Add docker cross-compile for arm64
This makes use of the docker buildx to cross-compile images for amd64/arm64. It is worth noting that there are ongoing issues with the Go compiler and qemu (used by buildx/buildkit) and the solution I ended up using here was to limit the affinity to `go build`. Better solutions may be forthcoming.

Refs:
https://github.com/golang/go/issues/24656
https://bugs.launchpad.net/qemu/+bug/1696773

This relates to issue #11 regarding container images built for running on Raspberry Pi.